Moving the Tensor Applier around
