+++ com.unity.ml-agents/Runtime/Training/TrainingForwardTensorApplier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2677467266ab48cfb01c5d873d043a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
