+++ com.unity.ml-agents/Runtime/Training/TrainingModelRunner.cs
@@ -23,12 +23,11 @@ namespace Unity.MLAgents
         ITensorAllocator m_TensorAllocator;
         TensorGenerator m_TensorGenerator;
         TrainingTensorGenerator m_TrainingTensorGenerator;
-        TensorApplier m_TensorApplier;
+        TrainingForwardTensorApplier m_TensorApplier;
 
         Model m_Model;
         IWorker m_Engine;
         bool m_Verbose = false;
-        string[] m_TrainingOutputNames;
         IReadOnlyList<TensorProxy> m_TrainingInputs;
         List<TensorProxy> m_TrainingOutputs;
         Dictionary<string, Tensor> m_InputsByName;
@@ -68,13 +67,13 @@ namespace Unity.MLAgents
             m_Engine = WorkerFactory.CreateWorker(executionDevice, barracudaModel, m_Verbose);
 
             m_TrainingInputs = barracudaModel.GetTrainingInputTensors();
-            m_TrainingOutputNames = barracudaModel.GetTrainingOutputNames();
+            InitializeTrainingState();
             m_TensorGenerator = new TensorGenerator(
                 seed, m_TensorAllocator, m_Memories, barracudaModel);
             m_TrainingTensorGenerator = new TrainingTensorGenerator(
                 seed, m_TensorAllocator, config.learningRate, config.gamma, barracudaModel);
-            m_TensorApplier = new TensorApplier(
-                actionSpec, seed, m_TensorAllocator, m_Memories, barracudaModel);
+            m_TensorApplier = new TrainingForwardTensorApplier(
+                actionSpec, seed, m_TensorAllocator, barracudaModel);
             m_InputsByName = new Dictionary<string, Tensor>();
             m_TrainingOutputs = new List<TensorProxy>();
             m_Buffer = buffer;
@@ -83,7 +82,8 @@ namespace Unity.MLAgents
         void InitializeTrainingState()
         {
             var initState = m_Model.GetTensorByName(TensorNames.InitialTrainingState);
-            m_TrainingState = new TensorProxy{
+            m_TrainingState = new TensorProxy
+            {
                 name = TensorNames.InitialTrainingState,
                 valueType = TensorProxy.TensorType.FloatingPoint,
                 data = initState,
@@ -188,7 +188,7 @@ namespace Unity.MLAgents
             // Execute the Model
             m_Engine.Execute(m_InputsByName);
 
-            FetchBarracudaOutputs(m_TrainingOutputNames);
+            FetchBarracudaOutputs(new string[] { TensorNames.TrainingOutput });
 
             // Update the outputs
             m_TensorApplier.ApplyTensors(m_TrainingOutputs, m_OrderedAgentsRequestingDecisions, m_LastActionsReceived);
@@ -213,10 +213,9 @@ namespace Unity.MLAgents
             // Execute the Model
             m_Engine.Execute(m_InputsByName);
 
-            FetchBarracudaOutputs(m_TrainingOutputNames);
-
             // Update the model
-            // m_TensorApplier.UpdateModel(m_TrainingOutputs, m_OrderedAgentsRequestingDecisions, m_LastActionsReceived);
+            FetchBarracudaOutputs(new string[] { TensorNames.TrainingStateOut });
+            m_TrainingState = m_TrainingOutputs[0];
         }
 
         public ActionBuffers GetAction(int agentId)
